Fix compilation warnings on macOS
https://ccpgames.atlassian.net/browse/PLAT-9355

+++ src/ResourceInfo/ResourceInfo.cpp
@@ -682,7 +682,7 @@ namespace CarbonResources
     	{
     		if( YAML::Node parameter = resource[m_binaryOperation.GetTag()] )
     		{
-    			m_binaryOperation = parameter.as<unsigned long>();
+    			m_binaryOperation = parameter.as<unsigned int>();
     		}
     		else
     		{
